Move eval fixtures out of root conftest to avoid heavy ML imports in unit test runs. Decreasing coverage requirement to 70% for unit tests.
